Refactor to avoid bundling DOMPurify into amp-bind and amp-mustache-0.1.
