+++ build-system/dep-check-config.js
@@ -49,8 +49,7 @@ exports.rules = [
     filesMatching: '**/*.js',
     mustNotDependOn: 'src/sanitizer.js',
     whitelist: [
-      // DEPRECATED! Do not extend this whitelist. Use src/purifier.js instead.
-      // Contact @choumx for questions.
+      // DEPRECATED: Use src/purifier.js instead. @choumx for questions.
       'extensions/amp-mustache/0.1/amp-mustache.js->src/sanitizer.js',
     ],
   },
@@ -58,8 +57,7 @@ exports.rules = [
     filesMatching: '**/*.js',
     mustNotDependOn: 'src/purifier.js',
     whitelist: [
-      'src/sanitizer.js->src/purifier.js',
-      'extensions/amp-bind/0.1/bind-impl.js->src/purifier.js',
+      // WARNING: Importing purifier.js will also bundle DOMPurify (13KB).
       'extensions/amp-mustache/0.2/amp-mustache.js->src/purifier.js',
       'extensions/amp-script/0.1/amp-script.js->src/purifier.js',
     ],
