+++ src/purifier.js
@@ -15,17 +15,17 @@
  */
 
 import {
-  checkCorsUrl,
-  getSourceUrl,
-  isProxyOrigin,
-  parseUrlDeprecated,
-  resolveRelativeUrl,
-} from './url';
-import {dict} from './utils/object';
-import {parseSrcset} from './srcset';
+  BIND_PREFIX,
+  BLACKLISTED_TAGS,
+  TRIPLE_MUSTACHE_WHITELISTED_TAGS,
+  WHITELISTED_ATTRS,
+  WHITELISTED_ATTRS_BY_TAGS,
+  WHITELISTED_TARGETS,
+  isValidAttr,
+} from './sanitation';
 import {remove} from './utils/array';
+import {rewriteAttributeValue} from './url-rewrite';
 import {startsWith} from './string';
-import {urls} from './config';
 import {user} from './log';
 import purify from 'dompurify/dist/purify.es';
 
@@ -39,136 +39,6 @@ const DomPurify = purify(self);
 
 /** @private @const {string} */
 const TAG = 'purifier';
-
-/** @private @const {string} */
-const ORIGINAL_TARGET_VALUE = '__AMP_ORIGINAL_TARGET_VALUE_';
-
-/** @private @const {string} */
-export const BIND_PREFIX = 'data-amp-bind-';
-
-/**
- * @const {!Object<string, boolean>}
- * See https://github.com/ampproject/amphtml/blob/master/spec/amp-html-format.md
- */
-export const BLACKLISTED_TAGS = {
-  'applet': true,
-  'audio': true,
-  'base': true,
-  'embed': true,
-  'frame': true,
-  'frameset': true,
-  'iframe': true,
-  'img': true,
-  'link': true,
-  'meta': true,
-  'object': true,
-  'style': true,
-  'video': true,
-};
-
-/**
- * Whitelist of tags allowed in triple mustache e.g. {{{name}}}.
- * Very restrictive by design since the triple mustache renders unescaped HTML
- * which, unlike double mustache, won't be processed by the AMP Validator.
- * @const {!Array<string>}
- */
-export const TRIPLE_MUSTACHE_WHITELISTED_TAGS = [
-  'a',
-  'b',
-  'br',
-  'caption',
-  'colgroup',
-  'code',
-  'del',
-  'div',
-  'em',
-  'i',
-  'ins',
-  'li',
-  'mark',
-  'ol',
-  'p',
-  'q',
-  's',
-  'small',
-  'span',
-  'strong',
-  'sub',
-  'sup',
-  'table',
-  'tbody',
-  'time',
-  'td',
-  'th',
-  'thead',
-  'tfoot',
-  'tr',
-  'u',
-  'ul',
-];
-
-/**
- * Tag-agnostic attribute whitelisted used by both Caja and DOMPurify.
- * @const {!Array<string>}
- */
-export const WHITELISTED_ATTRS = [
-  // AMP-only attributes that don't exist in HTML.
-  'amp-fx',
-  'fallback',
-  'heights',
-  'layout',
-  'min-font-size',
-  'max-font-size',
-  'on',
-  'option',
-  'placeholder',
-  // Attributes related to amp-form.
-  'submitting',
-  'submit-success',
-  'submit-error',
-  'validation-for',
-  'verify-error',
-  'visible-when-invalid',
-  // HTML attributes that are scrubbed by Caja but we handle specially.
-  'href',
-  'style',
-  // Attributes for amp-bind that exist in "[foo]" form.
-  'text',
-  // Attributes for amp-subscriptions.
-  'subscriptions-action',
-  'subscriptions-actions',
-  'subscriptions-decorate',
-  'subscriptions-dialog',
-  'subscriptions-display',
-  'subscriptions-section',
-  'subscriptions-service',
-];
-
-/**
- * Attributes that are only whitelisted for specific, non-AMP elements.
- * @const {!Object<string, !Array<string>>}
- */
-export const WHITELISTED_ATTRS_BY_TAGS = {
-  'a': [
-    'rel',
-    'target',
-  ],
-  'div': [
-    'template',
-  ],
-  'form': [
-    'action-xhr',
-    'verify-xhr',
-    'custom-validation-reporting',
-    'target',
-  ],
-  'input': [
-    'mask-output',
-  ],
-  'template': [
-    'type',
-  ],
-};
 
 /**
  * Tags that are only whitelisted for specific values of given attributes.
@@ -184,55 +54,6 @@ const WHITELISTED_TAGS_BY_ATTRS = {
   },
 };
 
-/** @const {!Array<string>} */
-export const WHITELISTED_TARGETS = ['_top', '_blank'];
-
-/** @const {!Array<string>} */
-const BLACKLISTED_ATTR_VALUES = [
-  /*eslint no-script-url: 0*/ 'javascript:',
-  /*eslint no-script-url: 0*/ 'vbscript:',
-  /*eslint no-script-url: 0*/ 'data:',
-  /*eslint no-script-url: 0*/ '<script',
-  /*eslint no-script-url: 0*/ '</script',
-];
-
-/** @const {!Object<string, !Object<string, !RegExp>>} */
-const BLACKLISTED_TAG_SPECIFIC_ATTR_VALUES = dict({
-  'input': {
-    'type': /(?:image|button)/i,
-  },
-});
-
-/** @const {!Array<string>} */
-const BLACKLISTED_FIELDS_ATTR = [
-  'form',
-  'formaction',
-  'formmethod',
-  'formtarget',
-  'formnovalidate',
-  'formenctype',
-];
-
-/** @const {!Object<string, !Array<string>>} */
-const BLACKLISTED_TAG_SPECIFIC_ATTRS = dict({
-  'input': BLACKLISTED_FIELDS_ATTR,
-  'textarea': BLACKLISTED_FIELDS_ATTR,
-  'select': BLACKLISTED_FIELDS_ATTR,
-});
-
-/**
- * Test for invalid `style` attribute values.
- *
- * !important avoids overriding AMP styles, while `position:fixed|sticky` is a
- * FixedLayer limitation (it only scans the style[amp-custom] stylesheet
- * for potential fixed/sticky elements). Note that the latter can be
- * circumvented with CSS comments -- not a big deal.
- *
- * @const {!RegExp}
- */
-const INVALID_INLINE_STYLE_REGEX =
-    /!important|position\s*:\s*fixed|position\s*:\s*sticky/i;
-
 const PURIFY_CONFIG = /** @type {!DomPurifyConfig} */ ({
   USE_PROFILES: {
     html: true,
@@ -240,7 +61,6 @@ const PURIFY_CONFIG = /** @type {!DomPurifyConfig} */ ({
     svgFilters: true,
   },
 });
-
 
 /**
  * Monotonically increasing counter used for keying nodes.
@@ -266,10 +86,11 @@ export function purifyHtml(dirty, diffing = false) {
  * Returns DOMPurify config for normal, escaped templates.
  * Do not use for unescaped templates.
  *
- * NOTE: see that we use DomPurifyConfig found in
+ * NOTE: See that we use DomPurifyConfig found in
  * build-system/dompurify.extern.js as the exact type. This is to prevent
  * closure compiler from optimizing these fields here in this file and in the
  * 3rd party library file. See #19624 for further information.
+ *
  * @return {!DomPurifyConfig}
  */
 export function purifyConfig() {
@@ -518,208 +339,4 @@ export function purifyTagsForTripleMustache(html, doc = self.document) {
   const div = doc.createElement('div');
   div.appendChild(fragment);
   return div./*OK*/innerHTML;
-}
-
-/**
- * Whether the attribute/value is valid.
- * @param {string} tagName Lowercase tag name.
- * @param {string} attrName Lowercase attribute name.
- * @param {string} attrValue
- * @param {boolean} opt_purify Is true, skips some attribute sanitizations
- *     that are already covered by DOMPurify.
- * @return {boolean}
- */
-export function isValidAttr(tagName, attrName, attrValue, opt_purify = false) {
-  if (!opt_purify) {
-    // "on*" attributes are not allowed.
-    if (startsWith(attrName, 'on') && attrName != 'on') {
-      return false;
-    }
-
-    // No attributes with "javascript" or other blacklisted substrings in them.
-    if (attrValue) {
-      const normalized = attrValue.toLowerCase().replace(/[\s,\u0000]+/g, '');
-      for (let i = 0; i < BLACKLISTED_ATTR_VALUES.length; i++) {
-        if (normalized.indexOf(BLACKLISTED_ATTR_VALUES[i]) >= 0) {
-          return false;
-        }
-      }
-    }
-  }
-
-  // Don't allow certain inline style values.
-  if (attrName == 'style') {
-    return !INVALID_INLINE_STYLE_REGEX.test(attrValue);
-  }
-
-  // Don't allow CSS class names with internal AMP prefix.
-  if (attrName == 'class' && attrValue && /(^|\W)i-amphtml-/i.test(attrValue)) {
-    return false;
-  }
-
-  // Don't allow '__amp_source_origin' in URLs.
-  if (isUrlAttribute(attrName) && /__amp_source_origin/.test(attrValue)) {
-    return false;
-  }
-
-  // Remove blacklisted attributes from specific tags e.g. input[formaction].
-  const attrNameBlacklist = BLACKLISTED_TAG_SPECIFIC_ATTRS[tagName];
-  if (attrNameBlacklist && attrNameBlacklist.indexOf(attrName) != -1) {
-    return false;
-  }
-
-  // Remove blacklisted values for specific attributes for specific tags
-  // e.g. input[type=image].
-  const attrBlacklist = BLACKLISTED_TAG_SPECIFIC_ATTR_VALUES[tagName];
-  if (attrBlacklist) {
-    const blacklistedValuesRegex = attrBlacklist[attrName];
-    if (blacklistedValuesRegex &&
-        attrValue.search(blacklistedValuesRegex) != -1) {
-      return false;
-    }
-  }
-
-  return true;
-}
-
-/**
- * The same as rewriteAttributeValue() but actually updates the element and
- * modifies other related attribute(s) for special cases, i.e. `target` for <a>.
- * @param {!Element} element
- * @param {string} attrName
- * @param {string} attrValue
- * @param {!Location=} opt_location
- * @param {boolean=} opt_updateProperty
- * @return {string}
- */
-export function rewriteAttributesForElement(
-  element, attrName, attrValue, opt_location, opt_updateProperty)
-{
-  const tag = element.tagName.toLowerCase();
-  const attr = attrName.toLowerCase();
-  const rewrittenValue = rewriteAttributeValue(tag, attr, attrValue);
-  // When served from proxy (CDN), changing an <a> tag from a hash link to a
-  // non-hash link requires updating `target` attribute per cache modification
-  // rules. @see amp-cache-modifications.md#url-rewrites
-  const isProxy = isProxyOrigin(opt_location || self.location);
-  if (isProxy && tag === 'a' && attr === 'href') {
-    const oldValue = element.getAttribute(attr);
-    const newValueIsHash = rewrittenValue[0] === '#';
-    const oldValueIsHash = oldValue && oldValue[0] === '#';
-
-    if (newValueIsHash && !oldValueIsHash) {
-      // Save the original value of `target` so it can be restored (if needed).
-      if (!element[ORIGINAL_TARGET_VALUE]) {
-        element[ORIGINAL_TARGET_VALUE] = element.getAttribute('target');
-      }
-      element.removeAttribute('target');
-    } else if (oldValueIsHash && !newValueIsHash) {
-      // Restore the original value of `target` or default to `_top`.
-      element.setAttribute('target', element[ORIGINAL_TARGET_VALUE] || '_top');
-    }
-  }
-  if (opt_updateProperty) {
-    // Must be done first for <input> elements to correctly update the UI for
-    // the first change on Safari and Chrome.
-    element[attr] = rewrittenValue;
-  }
-  element.setAttribute(attr, rewrittenValue);
-  return rewrittenValue;
-}
-
-/**
- * If (tagName, attrName) is a CDN-rewritable URL attribute, returns the
- * rewritten URL value. Otherwise, returns the unchanged `attrValue`.
- * See resolveUrlAttr() for rewriting rules.
- * @param {string} tagName Lowercase tag name.
- * @param {string} attrName Lowercase attribute name.
- * @param {string} attrValue
- * @return {string}
- * @private
- * @visibleForTesting
- */
-export function rewriteAttributeValue(tagName, attrName, attrValue) {
-  if (isUrlAttribute(attrName)) {
-    return resolveUrlAttr(tagName, attrName, attrValue, self.location);
-  }
-  return attrValue;
-}
-
-/**
- * @param {string} attrName Lowercase attribute name.
- * @return {boolean}
- */
-function isUrlAttribute(attrName) {
-  return (attrName == 'src' || attrName == 'href' || attrName == 'srcset');
-}
-
-/**
- * Rewrites the URL attribute values. URLs are rewritten as following:
- * - If URL is absolute, it is not rewritten
- * - If URL is relative, it's rewritten as absolute against the source origin
- * - If resulting URL is a `http:` URL and it's for image, the URL is rewritten
- *   again to be served with AMP Cache (cdn.ampproject.org).
- *
- * @param {string} tagName Lowercase tag name.
- * @param {string} attrName Lowercase attribute name.
- * @param {string} attrValue
- * @param {!Location} windowLocation
- * @return {string}
- * @private
- * @visibleForTesting
- */
-export function resolveUrlAttr(tagName, attrName, attrValue, windowLocation) {
-  checkCorsUrl(attrValue);
-  const isProxyHost = isProxyOrigin(windowLocation);
-  const baseUrl = parseUrlDeprecated(getSourceUrl(windowLocation));
-
-  if (attrName == 'href' && !startsWith(attrValue, '#')) {
-    return resolveRelativeUrl(attrValue, baseUrl);
-  }
-
-  if (attrName == 'src') {
-    if (tagName == 'amp-img') {
-      return resolveImageUrlAttr(attrValue, baseUrl, isProxyHost);
-    }
-    return resolveRelativeUrl(attrValue, baseUrl);
-  }
-
-  if (attrName == 'srcset') {
-    let srcset;
-    try {
-      srcset = parseSrcset(attrValue);
-    } catch (e) {
-      // Do not fail the whole template just because one srcset is broken.
-      // An AMP element will pick it up and report properly.
-      user().error(TAG, 'Failed to parse srcset: ', e);
-      return attrValue;
-    }
-    return srcset.stringify(url => resolveImageUrlAttr(url, baseUrl,
-        isProxyHost));
-  }
-
-  return attrValue;
-}
-
-/**
- * Non-HTTPs image URLs are rewritten via proxy.
- * @param {string} attrValue
- * @param {!Location} baseUrl
- * @param {boolean} isProxyHost
- * @return {string}
- */
-function resolveImageUrlAttr(attrValue, baseUrl, isProxyHost) {
-  const src = parseUrlDeprecated(resolveRelativeUrl(attrValue, baseUrl));
-
-  // URLs such as `data:` or proxy URLs are returned as is. Unsafe protocols
-  // do not arrive here - already stripped by the sanitizer.
-  if (src.protocol == 'data:' || isProxyOrigin(src) || !isProxyHost) {
-    return src.href;
-  }
-
-  // Rewrite as a proxy URL.
-  return `${urls.cdn}/i/` +
-      (src.protocol == 'https:' ? 's/' : '') +
-      encodeURIComponent(src.host) +
-      src.pathname + (src.search || '') + (src.hash || '');
 }
